Moved app initialization to Anodeutil
